feat: modularizo para no tener todo en el archivo admin
